Rename enum auth::Request into AuthRequest 😅
RD-47
